The CI gods were reading from a different conftest than the one I edited
