+++ conftest.py
@@ -10,23 +10,30 @@ def pytest_generate_tests(metafunc):
         metafunc.parametrize("device", [option_value])
 
 
-collect_ignore = ["setup.py", "speechbrain/integrations/"]
+collect_ignore = [
+    "setup.py",
+    "speechbrain/integrations/",
+    "speechbrain/lobes/models/fairseq_wav2vec.py",
+]
 try:
     import numba  # noqa: F401
 except ModuleNotFoundError:
     collect_ignore.append("speechbrain/nnet/loss/transducer_loss.py")
 try:
-    import kenlm  # noqa: F401
-except ModuleNotFoundError:
-    collect_ignore.append("speechbrain/decoders/language_model.py")
-try:
     import sklearn  # noqa: F401
 except ModuleNotFoundError:
     collect_ignore.append("speechbrain/utils/kmeans.py")
-    collect_ignore.append(
-        "speechbrain/integrations/huggingface/discrete_ssl.py"
-    )
 try:
     import sacrebleu  # noqa: F401
 except ModuleNotFoundError:
     collect_ignore.append("speechbrain/utils/bleu.py")
+try:
+    import speechtokenizer  # noqa: F401
+except ModuleNotFoundError:
+    collect_ignore.append(
+        "speechbrain/lobes/models/discrete/speechtokenizer_interface.py"
+    )
+try:
+    import ctc_segmentation  # noqa: F401
+except ModuleNotFoundError:
+    collect_ignore.append("speechbrain/alignment/ctc_segmentation.py")
